pyproject.toml
[project]
name = "poe-api-wrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "beautifulsoup4>=4.14.3",
    "cryptography>=43.0.3",
    "fastapi>=0.116.1",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.2",
    "orjson>=3.11.7",
    "pydantic>=2.11.7",
    "pymongo>=4.15.2",
    "pyexecjs>=1.5.1",
    "quickjs>=1.19.4",
    "requests-toolbelt>=1.0.0",
    "rich==13.3.4",
    "tiktoken>=0.12.0",
    "uvicorn>=0.35.0",
    "websocket>=0.2.1",
    "websocket-client>=1.9.0",
    "openai>=2.24.0",
    "tzdata>=2025.3",
    "python-multipart>=0.0.22",
    "curl-cffi>=0.14.0",
]
